feat(neovim): update keymaps and plugins
This commit updates the keymaps and plugins for Neovim. It introduces
new keybindings for Telescope, LSP, and other functionalities. It also
adds new plugins like 'trouble.nvim' and updates the configuration of
existing plugins like 'nvim-cmp' and 'nvim-treesitter'. The changes
improve the user experience by providing more functionalities and
making the editor more efficient to use.